refactor(frontend): apply FE review fixes (high/medium/low)
- [High] EditableEdge: replace inline outline:none + JS focus handlers
  with CSS module (.edgeLabelInput:focus rule)
- [Medium] AddNodeToolbar: replace JS onFocus/onBlur outline toggle
  with CSS module (.addNodeBtn:focus-visible rule)
- [Medium] graphConverter: replace `as GraphNodeType` assertion with
  Set-based narrowing via toNodeType() helper
- [Low] editableNodes.module.css: font-weight 400→330, letter-spacing 0.3px→0.54px
- [Low] Move graphToWorkflow.ts: ui/ → lib/ (FSD convention)
- [Low] EditableActionNode: aria-label "policyRef" → "정책 참조 코드"
