+++ frontend/src/features/update-workflow/ui/AddNodeToolbar.tsx
@@ -1,4 +1,5 @@
 import type { GraphNodeType } from "@/entities/workflow";
+import styles from "./addNodeToolbar.module.css";
 
 const NODE_TYPES: GraphNodeType[] = [
   "START",
@@ -15,38 +16,13 @@ interface AddNodeToolbarProps {
 
 export function AddNodeToolbar({ onAddNode }: AddNodeToolbarProps) {
   return (
-    <div
-      style={{
-        display: "flex",
-        gap: "6px",
-        padding: "8px",
-        borderBottom: "1px solid var(--text-primary, #000)",
-        flexWrap: "wrap",
-      }}
-    >
+    <div className={styles.toolbar}>
       {NODE_TYPES.map((type) => (
         <button
           key={type}
           type="button"
+          className={styles.addNodeBtn}
           onClick={() => onAddNode(type)}
-          style={{
-            background: "transparent",
-            border: "1px solid var(--text-primary, #000)",
-            borderRadius: "50px",
-            padding: "2px 10px",
-            fontSize: "11px",
-            fontFamily: "inherit",
-            letterSpacing: "-0.1px",
-            cursor: "pointer",
-            color: "var(--text-primary, #000)",
-          }}
-          onFocus={(e) => {
-            (e.currentTarget as HTMLButtonElement).style.outline =
-              "dashed 2px var(--text-primary, #000)";
-          }}
-          onBlur={(e) => {
-            (e.currentTarget as HTMLButtonElement).style.outline = "none";
-          }}
         >
           + {type}
         </button>
