+++ frontend/src/features/update-workflow/ui/addNodeToolbar.module.css
@@ -1,0 +1,25 @@
+.toolbar {
+  display: flex;
+  gap: 6px;
+  padding: 8px;
+  border-bottom: 1px solid var(--text-primary, #000);
+  flex-wrap: wrap;
+}
+
+.addNodeBtn {
+  background: transparent;
+  border: 1px solid var(--text-primary, #000);
+  border-radius: 50px;
+  padding: 2px 10px;
+  font-size: 11px;
+  font-family: inherit;
+  letter-spacing: -0.1px;
+  cursor: pointer;
+  color: var(--text-primary, #000);
+  outline: none;
+}
+
+.addNodeBtn:focus-visible {
+  outline: dashed 2px var(--text-primary, #000);
+  outline-offset: 2px;
+}
